move HighlightColors.ts to correct folder
