+++ www/src/components/diagram/CenterLine.tsx
@@ -7,7 +7,7 @@ import { useCommissioningPackageContext } from "../../hooks/useCommissioningPack
 import React from "react";
 import selectHandleFunction from "../../utils/HandlerFunctionHelper.tsx";
 import ToolContext from "../../context/ToolContext.ts";
-import HighlightColors from "../../types/HighlightColors.ts";
+import HighlightColors from "../../enums/HighlightColors.ts";
 import { iriFromSvgNode } from "../../utils/HelperFunctions.ts";
 
 interface CenterLineComponentProps {
