+++ www/src/enums/HighlightColors.ts
